health scripts and camera changes
Created health scripts for the player and enemies using a base class for
the health and subclasses for player and enemies. The camera to player
script was also changed to include more options for where the camera can
be changed and rotation to it, along with scrolling when the mouse is at
the edge.